optimize: テストファイルサイズ最適化 - 300行制限対応
Claude AIレビューコメント対応として、CLAUDE.md 300行制限ルールに適合するよう
大規模テストファイルを機能別に分割:

- test_conversion_state.py (639行) → basic/advanced分割
- test_file_model.py (697行) → basic/advanced分割
- test_dialogs.py (535行) → basic/advanced分割
- test_gui_models.py (455行) → basic/advanced分割
- test_validators.py (581行) → core/advanced/performance分割

全分割ファイルが300行制限を遵守し、機能別論理分割により
メンテナンス性も向上。
