+++ tests/test_conversion_state_advanced.py
@@ -1,0 +1,333 @@
+"""Conversion State Advanced Tests for Issue 501 Phase 4C
+
+This module provides advanced tests for ConversionState class
+to ensure proper timing, callbacks, and thread safety.
+"""
+
+import threading
+import time
+from unittest.mock import MagicMock, Mock, patch
+
+import pytest
+
+from kumihan_formatter.gui_models.conversion_state import ConversionState
+from tests.test_base import BaseTestCase
+
+
+class TestConversionStateTiming(BaseTestCase):
+    """Test ConversionState timing functionality"""
+
+    def test_timing_start(self):
+        """Test starting timing"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test timing start
+            if hasattr(conversion_state, "start_timing"):
+                conversion_state.start_timing()
+
+                if hasattr(conversion_state, "get_start_time"):
+                    start_time = conversion_state.get_start_time()
+                    assert start_time is not None
+                elif hasattr(conversion_state, "start_time"):
+                    assert conversion_state.start_time is not None
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+    def test_timing_end(self):
+        """Test ending timing"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test timing end
+            if hasattr(conversion_state, "start_timing"):
+                conversion_state.start_timing()
+                time.sleep(0.1)  # Small delay
+
+                if hasattr(conversion_state, "end_timing"):
+                    conversion_state.end_timing()
+
+                    if hasattr(conversion_state, "get_end_time"):
+                        end_time = conversion_state.get_end_time()
+                        assert end_time is not None
+                    elif hasattr(conversion_state, "end_time"):
+                        assert conversion_state.end_time is not None
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+    def test_duration_calculation(self):
+        """Test duration calculation"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test duration calculation
+            if hasattr(conversion_state, "start_timing"):
+                conversion_state.start_timing()
+                time.sleep(0.1)  # Small delay
+
+                if hasattr(conversion_state, "end_timing"):
+                    conversion_state.end_timing()
+
+                    if hasattr(conversion_state, "get_duration"):
+                        duration = conversion_state.get_duration()
+                        assert duration is not None
+                        assert duration >= 0.1  # Should be at least the sleep time
+                    elif hasattr(conversion_state, "duration"):
+                        assert conversion_state.duration is not None
+                        assert conversion_state.duration >= 0.1
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+
+class TestConversionStateCallbacks(BaseTestCase):
+    """Test ConversionState callback functionality"""
+
+    def test_status_callbacks(self):
+        """Test status change callbacks"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test status callbacks
+            callback_called = False
+            callback_status = None
+
+            def status_callback(status):
+                nonlocal callback_called, callback_status
+                callback_called = True
+                callback_status = status
+
+            # Test callback registration
+            if hasattr(conversion_state, "add_status_callback"):
+                conversion_state.add_status_callback(status_callback)
+
+                if hasattr(conversion_state, "set_status"):
+                    conversion_state.set_status("converting")
+                    assert callback_called
+                    assert callback_status == "converting"
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+    def test_completion_callbacks(self):
+        """Test completion callbacks"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test completion callbacks
+            callback_called = False
+
+            def completion_callback():
+                nonlocal callback_called
+                callback_called = True
+
+            # Test callback registration
+            if hasattr(conversion_state, "add_completion_callback"):
+                conversion_state.add_completion_callback(completion_callback)
+
+                if hasattr(conversion_state, "set_status"):
+                    conversion_state.set_status("completed")
+                    assert callback_called
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+    def test_error_callbacks(self):
+        """Test error callbacks"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test error callbacks
+            callback_called = False
+            callback_error = None
+
+            def error_callback(error):
+                nonlocal callback_called, callback_error
+                callback_called = True
+                callback_error = error
+
+            # Test callback registration
+            if hasattr(conversion_state, "add_error_callback"):
+                conversion_state.add_error_callback(error_callback)
+
+                if hasattr(conversion_state, "set_error"):
+                    conversion_state.set_error("Test error")
+                    assert callback_called
+                    assert callback_error == "Test error"
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+
+class TestConversionStateThreadSafety(BaseTestCase):
+    """Test ConversionState thread safety"""
+
+    def test_concurrent_status_updates(self):
+        """Test concurrent status updates"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test concurrent updates
+            results = []
+
+            def update_status(status):
+                if hasattr(conversion_state, "set_status"):
+                    conversion_state.set_status(status)
+                    results.append(status)
+
+            # Create multiple threads
+            threads = []
+            statuses = ["preparing", "converting", "completed"]
+
+            for status in statuses:
+                thread = threading.Thread(target=update_status, args=(status,))
+                threads.append(thread)
+                thread.start()
+
+            # Wait for all threads
+            for thread in threads:
+                thread.join()
+
+            # All updates should complete
+            assert len(results) == len(statuses)
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+    def test_concurrent_progress_updates(self):
+        """Test concurrent progress updates"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test concurrent progress updates
+            results = []
+
+            def update_progress(progress):
+                if hasattr(conversion_state, "set_progress"):
+                    conversion_state.set_progress(progress)
+                    results.append(progress)
+
+            # Create multiple threads
+            threads = []
+            progress_values = [25, 50, 75, 100]
+
+            for progress in progress_values:
+                thread = threading.Thread(target=update_progress, args=(progress,))
+                threads.append(thread)
+                thread.start()
+
+            # Wait for all threads
+            for thread in threads:
+                thread.join()
+
+            # All updates should complete
+            assert len(results) == len(progress_values)
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+
+class TestConversionStateIntegration(BaseTestCase):
+    """Test ConversionState integration scenarios"""
+
+    def test_conversion_workflow(self):
+        """Test complete conversion workflow"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test complete workflow
+            workflow_steps = [
+                ("idle", 0),
+                ("preparing", 0),
+                ("converting", 25),
+                ("converting", 50),
+                ("converting", 75),
+                ("completed", 100),
+            ]
+
+            for status, progress in workflow_steps:
+                if hasattr(conversion_state, "set_status"):
+                    conversion_state.set_status(status)
+                if hasattr(conversion_state, "set_progress"):
+                    conversion_state.set_progress(progress)
+
+                # Verify state consistency
+                if hasattr(conversion_state, "get_status"):
+                    assert conversion_state.get_status() == status
+                if hasattr(conversion_state, "get_progress"):
+                    assert conversion_state.get_progress() == progress
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+    def test_error_recovery_workflow(self):
+        """Test error recovery workflow"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test error recovery
+            if hasattr(conversion_state, "set_status"):
+                conversion_state.set_status("converting")
+                conversion_state.set_status("error")
+
+                # Recovery
+                conversion_state.set_status("idle")
+                assert conversion_state.get_status() == "idle"
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+    def test_cancellation_workflow(self):
+        """Test cancellation workflow"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test cancellation
+            if hasattr(conversion_state, "set_status"):
+                conversion_state.set_status("converting")
+                conversion_state.set_status("cancelled")
+
+                assert conversion_state.get_status() == "cancelled"
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+
+class TestConversionStatePerformance(BaseTestCase):
+    """Test ConversionState performance"""
+
+    def test_rapid_updates(self):
+        """Test rapid state updates"""
+        try:
+            conversion_state = ConversionState()
+
+            # Test rapid updates
+            for i in range(100):
+                if hasattr(conversion_state, "set_progress"):
+                    conversion_state.set_progress(i)
+
+            # Should handle rapid updates without issues
+            assert conversion_state is not None
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
+
+    def test_memory_efficiency(self):
+        """Test memory efficiency"""
+        try:
+            # Create multiple state objects
+            states = []
+            for i in range(50):
+                state = ConversionState()
+                states.append(state)
+
+            # Should not cause memory issues
+            assert len(states) == 50
+
+            # Cleanup
+            del states
+
+        except ImportError:
+            pytest.skip("ConversionState not available")
