+++ tests/test_dialogs_basic.py
@@ -308,228 +308,3 @@ class TestCustomDialogs(BaseTestCase):
 
 class TestDialogInteraction(BaseTestCase):
     """Test dialog user interaction"""
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_button_interaction(self, mock_toplevel):
-        """Test dialog button interactions"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-
-        # Mock button interactions
-        with patch("tkinter.Button") as mock_button:
-            mock_button_instance = Mock()
-            mock_button.return_value = mock_button_instance
-
-            # Test button callbacks
-            callback_called = False
-
-            def test_callback():
-                nonlocal callback_called
-                callback_called = True
-
-            # Simulate button click
-            try:
-                test_callback()
-                assert callback_called
-
-            except ImportError:
-                pytest.skip("Dialog button interaction not available")
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_data_validation(self, mock_toplevel):
-        """Test dialog data validation"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-
-        # Mock form validation
-        with patch("tkinter.Entry") as mock_entry:
-            mock_entry_instance = Mock()
-            mock_entry.return_value = mock_entry_instance
-
-            # Test validation scenarios
-            test_inputs = [
-                ("valid_input", True),
-                ("", False),
-                ("123", True),
-                ("invalid@#$", False),
-            ]
-
-            for input_value, expected_valid in test_inputs:
-                mock_entry_instance.get.return_value = input_value
-                # Validation logic testing
-                assert isinstance(expected_valid, bool)
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_keyboard_shortcuts(self, mock_toplevel):
-        """Test dialog keyboard shortcuts"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-        mock_dialog.bind = Mock()
-
-        # Test keyboard shortcuts
-        shortcuts = [
-            ("<Return>", "ok_action"),
-            ("<Escape>", "cancel_action"),
-            ("<Control-s>", "save_action"),
-        ]
-
-        for shortcut, action in shortcuts:
-            try:
-                # Test keyboard shortcut binding
-                assert shortcut and action
-
-            except ImportError:
-                pytest.skip("Dialog keyboard shortcuts not available")
-
-
-class TestDialogErrorHandling(BaseTestCase):
-    """Test dialog error handling"""
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_creation_error(self, mock_toplevel):
-        """Test dialog creation error handling"""
-        mock_toplevel.side_effect = Exception("Dialog creation failed")
-
-        with pytest.raises(Exception) as exc_info:
-            # Test error handling
-            mock_toplevel()
-
-        assert "Dialog creation failed" in str(exc_info.value)
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_invalid_input_handling(self, mock_toplevel):
-        """Test handling of invalid input in dialogs"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-
-        # Test invalid input scenarios
-        invalid_inputs = [
-            None,
-            "",
-            "   ",  # Whitespace only
-            "a" * 1000,  # Too long
-        ]
-
-        for invalid_input in invalid_inputs:
-            # Invalid input testing
-            assert invalid_input is None or isinstance(invalid_input, str)
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_resource_cleanup(self, mock_toplevel):
-        """Test dialog resource cleanup"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-        mock_dialog.destroy = Mock()
-
-        try:
-            # Test dialog cleanup
-            mock_dialog.destroy()
-            assert mock_dialog.destroy.called
-
-        except ImportError:
-            pytest.skip("Dialog resource cleanup not available")
-
-
-class TestDialogAccessibility(BaseTestCase):
-    """Test dialog accessibility features"""
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_focus_management(self, mock_toplevel):
-        """Test dialog focus management"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-        mock_dialog.focus_set = Mock()
-
-        # Test initial focus
-        try:
-            mock_dialog.focus_set()
-            assert mock_dialog.focus_set.called
-
-        except ImportError:
-            pytest.skip("Dialog focus management not available")
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_tab_navigation(self, mock_toplevel):
-        """Test dialog tab navigation"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-
-        # Mock tab navigation
-        with patch("tkinter.Entry") as mock_entry:
-            with patch("tkinter.Button") as mock_button:
-                # Test tab order
-                try:
-                    # Tab navigation testing
-                    assert mock_entry.called or mock_button.called or True
-
-                except ImportError:
-                    pytest.skip("Dialog tab navigation not available")
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_screen_reader_support(self, mock_toplevel):
-        """Test dialog screen reader support"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-
-        # Test accessibility attributes
-        accessibility_attrs = [
-            "title",
-            "description",
-            "role",
-        ]
-
-        for attr in accessibility_attrs:
-            # Accessibility testing
-            assert attr and isinstance(attr, str)
-
-
-class TestDialogPerformance(BaseTestCase):
-    """Test dialog performance characteristics"""
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_creation_speed(self, mock_toplevel):
-        """Test dialog creation performance"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-
-        # Test multiple dialog creation
-        for i in range(10):
-            try:
-                dialog = mock_toplevel()
-                assert dialog is not None
-
-            except ImportError:
-                pytest.skip("Dialog creation speed not available")
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_memory_usage(self, mock_toplevel):
-        """Test dialog memory usage"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-
-        # Test memory usage with multiple dialogs
-        dialogs = []
-        for i in range(5):
-            dialog = mock_toplevel()
-            dialogs.append(dialog)
-
-        # Should not cause memory issues
-        assert len(dialogs) == 5
-
-        # Cleanup
-        del dialogs
-
-    @patch("tkinter.Toplevel")
-    def test_dialog_responsiveness(self, mock_toplevel):
-        """Test dialog responsiveness"""
-        mock_dialog = Mock()
-        mock_toplevel.return_value = mock_dialog
-        mock_dialog.update = Mock()
-
-        # Test dialog responsiveness
-        try:
-            mock_dialog.update()
-            assert mock_dialog.update.called
-
-        except ImportError:
-            pytest.skip("Dialog responsiveness not available")
