+++ tests/test_file_model_basic.py
@@ -1,0 +1,320 @@
+"""File Model Basic Tests for Issue 501 Phase 4C
+
+This module provides basic tests for FileModel class
+to ensure proper file operations and validation.
+"""
+
+import os
+import tempfile
+from pathlib import Path
+from unittest.mock import MagicMock, Mock, patch
+
+import pytest
+
+from kumihan_formatter.gui_models.file_model import FileModel
+from tests.test_base import BaseTestCase
+
+
+class TestFileModelBasics(BaseTestCase):
+    """Test basic FileModel functionality"""
+
+    def test_file_model_initialization(self):
+        """Test FileModel initialization"""
+        try:
+            file_model = FileModel()
+            assert file_model is not None
+
+            # Test initial state
+            if hasattr(file_model, "file_path"):
+                assert file_model.file_path is None or file_model.file_path == ""
+            if hasattr(file_model, "file_info"):
+                assert file_model.file_info is None or isinstance(
+                    file_model.file_info, dict
+                )
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+    def test_file_model_default_values(self):
+        """Test default values after initialization"""
+        try:
+            file_model = FileModel()
+
+            # Test default values
+            default_checks = [
+                ("file_path", [None, "", False]),
+                ("file_size", [None, 0, -1]),
+                ("file_extension", [None, "", False]),
+                ("encoding", [None, "", "utf-8"]),
+                ("is_valid", [None, False, True]),
+            ]
+
+            for attr, valid_values in default_checks:
+                if hasattr(file_model, attr):
+                    value = getattr(file_model, attr)
+                    assert value in valid_values or value is None
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+    def test_file_model_attributes(self):
+        """Test FileModel attributes"""
+        try:
+            file_model = FileModel()
+
+            # Test common attributes
+            common_attrs = [
+                "file_path",
+                "file_size",
+                "file_extension",
+                "file_name",
+                "directory",
+                "encoding",
+                "is_valid",
+                "last_modified",
+            ]
+
+            for attr in common_attrs:
+                # Attribute existence test
+                has_attr = hasattr(file_model, attr)
+                has_getter = hasattr(file_model, f"get_{attr}")
+                has_setter = hasattr(file_model, f"set_{attr}")
+
+                # At least one access method should exist
+                assert has_attr or has_getter or has_setter
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+
+class TestFileModelFileOperations(BaseTestCase):
+    """Test FileModel file operations"""
+
+    def test_file_path_setting(self):
+        """Test setting file path"""
+        try:
+            file_model = FileModel()
+
+            # Test file path setting
+            test_file = self.create_temp_file("test content", suffix=".txt")
+
+            if hasattr(file_model, "set_file_path"):
+                file_model.set_file_path(test_file)
+                current_path = file_model.get_file_path()
+                assert current_path == test_file
+            elif hasattr(file_model, "file_path"):
+                file_model.file_path = test_file
+                assert file_model.file_path == test_file
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+    def test_file_info_extraction(self):
+        """Test file information extraction"""
+        try:
+            file_model = FileModel()
+
+            # Test file info extraction
+            test_content = "テストファイル内容"
+            test_file = self.create_temp_file(test_content, suffix=".txt")
+
+            if hasattr(file_model, "set_file_path"):
+                file_model.set_file_path(test_file)
+
+                # Test file size
+                if hasattr(file_model, "get_file_size"):
+                    size = file_model.get_file_size()
+                    assert size > 0
+
+                # Test file extension
+                if hasattr(file_model, "get_file_extension"):
+                    ext = file_model.get_file_extension()
+                    assert ext == ".txt"
+
+                # Test file name
+                if hasattr(file_model, "get_file_name"):
+                    name = file_model.get_file_name()
+                    assert name.endswith(".txt")
+
+                # Test directory
+                if hasattr(file_model, "get_directory"):
+                    directory = file_model.get_directory()
+                    assert directory and os.path.isdir(directory)
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+    def test_file_existence_check(self):
+        """Test file existence checking"""
+        try:
+            file_model = FileModel()
+
+            # Test with existing file
+            test_file = self.create_temp_file("test content", suffix=".txt")
+
+            if hasattr(file_model, "set_file_path"):
+                file_model.set_file_path(test_file)
+
+                if hasattr(file_model, "file_exists"):
+                    assert file_model.file_exists() is True
+                elif hasattr(file_model, "is_valid"):
+                    assert file_model.is_valid is True
+
+            # Test with non-existent file
+            if hasattr(file_model, "set_file_path"):
+                file_model.set_file_path("/nonexistent/file.txt")
+
+                if hasattr(file_model, "file_exists"):
+                    assert file_model.file_exists() is False
+                elif hasattr(file_model, "is_valid"):
+                    assert file_model.is_valid is False
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+
+class TestFileModelValidation(BaseTestCase):
+    """Test FileModel validation functionality"""
+
+    def test_file_validation(self):
+        """Test file validation"""
+        try:
+            file_model = FileModel()
+
+            # Test valid files
+            valid_files = [
+                (self.create_temp_file("content", suffix=".txt"), True),
+                (self.create_temp_file("content", suffix=".md"), True),
+                (self.create_temp_file("content", suffix=".html"), True),
+            ]
+
+            for file_path, expected_valid in valid_files:
+                if hasattr(file_model, "validate_file"):
+                    is_valid = file_model.validate_file(file_path)
+                    assert is_valid == expected_valid
+                elif hasattr(file_model, "is_valid_file"):
+                    is_valid = file_model.is_valid_file(file_path)
+                    assert is_valid == expected_valid
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+    def test_file_format_validation(self):
+        """Test file format validation"""
+        try:
+            file_model = FileModel()
+
+            # Test different file formats
+            format_tests = [
+                (".txt", True),
+                (".md", True),
+                (".html", True),
+                (".doc", False),  # Potentially unsupported
+                (".pdf", False),  # Potentially unsupported
+            ]
+
+            for extension, expected_valid in format_tests:
+                test_file = self.create_temp_file("content", suffix=extension)
+
+                if hasattr(file_model, "validate_format"):
+                    is_valid = file_model.validate_format(test_file)
+                    # Note: Some formats might be supported
+                    assert isinstance(is_valid, bool)
+                elif hasattr(file_model, "is_supported_format"):
+                    is_valid = file_model.is_supported_format(extension)
+                    assert isinstance(is_valid, bool)
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+    def test_file_size_validation(self):
+        """Test file size validation"""
+        try:
+            file_model = FileModel()
+
+            # Test different file sizes
+            size_tests = [
+                ("small content", True),
+                ("medium content" * 100, True),
+                ("large content" * 10000, True),  # Large but manageable
+            ]
+
+            for content, expected_valid in size_tests:
+                test_file = self.create_temp_file(content, suffix=".txt")
+
+                if hasattr(file_model, "validate_size"):
+                    is_valid = file_model.validate_size(test_file)
+                    assert isinstance(is_valid, bool)
+                elif hasattr(file_model, "is_size_acceptable"):
+                    is_valid = file_model.is_size_acceptable(test_file)
+                    assert isinstance(is_valid, bool)
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+
+class TestFileModelEncoding(BaseTestCase):
+    """Test FileModel encoding functionality"""
+
+    def test_encoding_detection(self):
+        """Test encoding detection"""
+        try:
+            file_model = FileModel()
+
+            # Test UTF-8 encoding
+            test_content = "日本語テスト内容"
+            utf8_file = self.create_temp_file(test_content, suffix=".txt")
+
+            if hasattr(file_model, "detect_encoding"):
+                encoding = file_model.detect_encoding(utf8_file)
+                assert encoding in ["utf-8", "utf-8-sig", "UTF-8"]
+            elif hasattr(file_model, "get_encoding"):
+                encoding = file_model.get_encoding(utf8_file)
+                assert encoding is not None
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+    def test_encoding_validation(self):
+        """Test encoding validation"""
+        try:
+            file_model = FileModel()
+
+            # Test different encodings
+            encoding_tests = [
+                "utf-8",
+                "utf-8-sig",
+                "shift_jis",
+                "euc-jp",
+                "iso-8859-1",
+            ]
+
+            for encoding in encoding_tests:
+                if hasattr(file_model, "validate_encoding"):
+                    is_valid = file_model.validate_encoding(encoding)
+                    assert isinstance(is_valid, bool)
+                elif hasattr(file_model, "is_supported_encoding"):
+                    is_valid = file_model.is_supported_encoding(encoding)
+                    assert isinstance(is_valid, bool)
+
+        except ImportError:
+            pytest.skip("FileModel not available")
+
+    def test_encoding_conversion(self):
+        """Test encoding conversion"""
+        try:
+            file_model = FileModel()
+
+            # Test encoding conversion
+            test_content = "日本語コンテンツ"
+            test_file = self.create_temp_file(test_content, suffix=".txt")
+
+            if hasattr(file_model, "convert_encoding"):
+                result = file_model.convert_encoding(test_file, "utf-8")
+                assert result is not None
+            elif hasattr(file_model, "ensure_encoding"):
+                result = file_model.ensure_encoding(test_file, "utf-8")
+                assert result is not None
+
+        except ImportError:
+            pytest.skip("FileModel not available")
