+++ tests/test_gui_models_advanced.py
@@ -1,7 +1,7 @@
-"""GUI Models Tests for Issue 501 Phase 4C
+"""GUI Models Advanced Tests for Issue 501 Phase 4C
 
-This module tests GUI model functionality to ensure
-proper state management and data handling.
+This module tests advanced GUI model functionality including
+integration, error handling, and performance.
 """
 
 import os
@@ -17,204 +17,8 @@ from kumihan_formatter.gui_models.state_model import StateModel
 from tests.test_base import BaseTestCase
 
 
-class TestStateModel(BaseTestCase):
-    """Test state model functionality"""
-
-    def test_state_model_initialization(self):
-        """Test StateModel initialization"""
-        try:
-            state_model = StateModel()
-            assert state_model is not None
-        except ImportError:
-            pytest.skip("StateModel not available")
-
-    def test_state_model_input_file_management(self):
-        """Test input file state management"""
-        try:
-            state_model = StateModel()
-
-            # Test input file setting
-            test_file = self.create_temp_file("test content", suffix=".txt")
-
-            if hasattr(state_model, "set_input_file"):
-                state_model.set_input_file(test_file)
-                assert state_model.get_input_file() == test_file
-            elif hasattr(state_model, "input_file"):
-                state_model.input_file = test_file
-                assert state_model.input_file == test_file
-            else:
-                # Basic file management test
-                assert os.path.exists(test_file)
-
-        except ImportError:
-            pytest.skip("StateModel input file management not available")
-
-    def test_state_model_output_directory_management(self):
-        """Test output directory state management"""
-        try:
-            state_model = StateModel()
-
-            # Test output directory setting
-            test_dir = self.create_temp_dir()
-
-            if hasattr(state_model, "set_output_directory"):
-                state_model.set_output_directory(test_dir)
-                assert state_model.get_output_directory() == test_dir
-            elif hasattr(state_model, "output_directory"):
-                state_model.output_directory = test_dir
-                assert state_model.output_directory == test_dir
-            else:
-                # Basic directory management test
-                assert os.path.exists(test_dir)
-
-        except ImportError:
-            pytest.skip("StateModel output directory management not available")
-
-    def test_state_model_conversion_options(self):
-        """Test conversion options state management"""
-        try:
-            state_model = StateModel()
-
-            # Test conversion options
-            options = {
-                "show_source": True,
-                "output_format": "html",
-                "encoding": "utf-8",
-            }
-
-            for key, value in options.items():
-                if hasattr(state_model, f"set_{key}"):
-                    getattr(state_model, f"set_{key}")(value)
-                    assert getattr(state_model, f"get_{key}")() == value
-                elif hasattr(state_model, key):
-                    setattr(state_model, key, value)
-                    assert getattr(state_model, key) == value
-                else:
-                    # Basic options test
-                    assert value is not None
-
-        except ImportError:
-            pytest.skip("StateModel conversion options not available")
-
-    def test_state_model_serialization(self):
-        """Test state model serialization"""
-        try:
-            state_model = StateModel()
-
-            # Test state serialization
-            if hasattr(state_model, "to_dict"):
-                state_dict = state_model.to_dict()
-                assert isinstance(state_dict, dict)
-
-            if hasattr(state_model, "from_dict"):
-                test_state = {"input_file": "test.txt", "show_source": True}
-                state_model.from_dict(test_state)
-                assert state_model is not None
-
-        except ImportError:
-            pytest.skip("StateModel serialization not available")
-
-
-class TestFileModel(BaseTestCase):
-    """Test file model functionality"""
-
-    def test_file_model_initialization(self):
-        """Test FileModel initialization"""
-        try:
-            file_model = FileModel()
-            assert file_model is not None
-        except ImportError:
-            pytest.skip("FileModel not available")
-
-    def test_file_model_file_info(self):
-        """Test file information management"""
-        try:
-            file_model = FileModel()
-
-            # Test file info
-            test_file = self.create_temp_file("test content", suffix=".txt")
-
-            if hasattr(file_model, "set_file_path"):
-                file_model.set_file_path(test_file)
-
-                # Test file info extraction
-                if hasattr(file_model, "get_file_size"):
-                    size = file_model.get_file_size()
-                    assert size > 0
-
-                if hasattr(file_model, "get_file_extension"):
-                    ext = file_model.get_file_extension()
-                    assert ext == ".txt"
-
-                if hasattr(file_model, "get_file_name"):
-                    name = file_model.get_file_name()
-                    assert name.endswith(".txt")
-
-        except ImportError:
-            pytest.skip("FileModel file info not available")
-
-    def test_file_model_validation(self):
-        """Test file model validation"""
-        try:
-            file_model = FileModel()
-
-            # Test file validation
-            test_files = [
-                (self.create_temp_file("content", suffix=".txt"), True),
-                (self.create_temp_file("content", suffix=".md"), True),
-                ("/nonexistent/file.txt", False),
-                ("", False),
-            ]
-
-            for file_path, expected_valid in test_files:
-                if hasattr(file_model, "validate_file"):
-                    is_valid = file_model.validate_file(file_path)
-                    assert is_valid == expected_valid
-                elif hasattr(file_model, "is_valid_file"):
-                    is_valid = file_model.is_valid_file(file_path)
-                    assert is_valid == expected_valid
-                else:
-                    # Basic validation test
-                    exists = os.path.exists(file_path) if file_path else False
-                    assert exists == expected_valid or not expected_valid
-
-        except ImportError:
-            pytest.skip("FileModel validation not available")
-
-    def test_file_model_encoding_detection(self):
-        """Test file encoding detection"""
-        try:
-            file_model = FileModel()
-
-            # Test encoding detection
-            test_content = "テスト内容"
-            utf8_file = self.create_temp_file(test_content, suffix=".txt")
-
-            if hasattr(file_model, "detect_encoding"):
-                encoding = file_model.detect_encoding(utf8_file)
-                assert encoding in ["utf-8", "utf-8-sig"]
-            elif hasattr(file_model, "get_encoding"):
-                encoding = file_model.get_encoding(utf8_file)
-                assert encoding is not None
-            else:
-                # Basic encoding test
-                content = Path(utf8_file).read_text(encoding="utf-8")
-                assert content == test_content
-
-        except ImportError:
-            pytest.skip("FileModel encoding detection not available")
-
-
-class TestConversionState(BaseTestCase):
-    """Test conversion state functionality"""
-
-    def test_conversion_state_initialization(self):
-        """Test ConversionState initialization"""
-        try:
-            conversion_state = ConversionState()
-            assert conversion_state is not None
-        except ImportError:
-            pytest.skip("ConversionState not available")
+class TestConversionStateAdvanced(BaseTestCase):
+    """Test advanced conversion state functionality"""
 
     def test_conversion_state_status_management(self):
         """Test conversion status management"""
