--- tests/test_validators.py
@@ -1,581 +1,0 @@
-"""Validators Tests for Issue 500 Phase 3C
-
-This module tests validation functionality to ensure robust validation
-under various scenarios.
-"""
-
-import os
-import tempfile
-from pathlib import Path
-from unittest.mock import Mock, patch
-
-import pytest
-
-from kumihan_formatter.core.validators import (
-    DocumentValidator,
-    ValidationIssue,
-    ValidationReporter,
-)
-from kumihan_formatter.core.validators.performance_validator import PerformanceValidator
-from kumihan_formatter.core.validators.structure_validator import StructureValidator
-from kumihan_formatter.core.validators.syntax_validator import SyntaxValidator
-from tests.test_base import BaseTestCase, ValidatorTestCase, create_test_kumihan_content
-
-
-class TestDocumentValidator(ValidatorTestCase):
-    """Test document validator functionality"""
-
-    def test_document_validator_initialization(self):
-        """Test DocumentValidator initialization"""
-        self.test_component_initialization(DocumentValidator, "DocumentValidator")
-
-    def test_document_validator_basic_validation(self):
-        """Test basic document validation"""
-        try:
-            validator = DocumentValidator()
-
-            # Test basic validation with Kumihan content
-            test_content = ";;;強調;;; テスト内容 ;;;"
-            temp_file = self.create_temp_file(test_content)
-
-            # Test validation
-            if hasattr(validator, "validate"):
-                result = validator.validate(test_content)
-                assert result is not None
-                # Check if validation passed
-                assert hasattr(result, "is_valid") or isinstance(result, bool)
-
-            # Test file validation
-            if hasattr(validator, "validate_file"):
-                result = validator.validate_file(temp_file)
-                assert result is not None
-
-        except ImportError:
-            pytest.skip("DocumentValidator not available")
-
-    def test_document_validator_kumihan_syntax(self):
-        """Test Kumihan syntax validation"""
-        try:
-            validator = DocumentValidator()
-
-            # Test complex Kumihan syntax
-            test_content = create_test_kumihan_content()
-            temp_file = self.create_temp_file(test_content)
-
-            # Test validation of complex syntax
-            if hasattr(validator, "validate"):
-                result = validator.validate(test_content)
-                assert result is not None
-
-                # Check if validation recognizes Kumihan syntax
-                if hasattr(result, "notations_found"):
-                    assert len(result.notations_found) > 0
-                elif hasattr(result, "syntax_elements"):
-                    assert len(result.syntax_elements) > 0
-
-        except ImportError:
-            pytest.skip("DocumentValidator not available")
-
-    def test_document_validator_error_detection(self):
-        """Test error detection in documents"""
-        try:
-            validator = DocumentValidator()
-
-            # Test malformed Kumihan syntax
-            malformed_content = ";;;強調;;; 閉じタグなし"  # Missing closing tag
-            temp_file = self.create_temp_file(malformed_content)
-
-            # Test error detection
-            if hasattr(validator, "validate"):
-                result = validator.validate(malformed_content)
-                assert result is not None
-
-                # Check if errors were detected
-                if hasattr(result, "errors"):
-                    assert len(result.errors) > 0
-                elif hasattr(result, "is_valid"):
-                    # Should detect syntax error
-                    assert result.is_valid is False
-                elif isinstance(result, bool):
-                    # Should return False for invalid syntax
-                    assert result is False
-
-        except ImportError:
-            pytest.skip("DocumentValidator not available")
-
-
-class TestValidationIssue(BaseTestCase):
-    """Test validation issue functionality"""
-
-    def test_validation_issue_initialization(self):
-        """Test ValidationIssue initialization"""
-        try:
-            # Test with required parameters (both level and severity for compatibility)
-            issue = ValidationIssue(
-                level="warning",
-                category="syntax",
-                message="Test validation issue",
-                line_number=1,
-                column_number=1,
-            )
-            assert issue is not None
-            assert issue.message == "Test validation issue"
-            assert issue.level == "warning"
-            assert issue.category == "syntax"
-            assert issue.line_number == 1
-            assert issue.column_number == 1
-        except (ImportError, TypeError):
-            # If ValidationIssue requires different parameters, test basic creation
-            try:
-                issue = ValidationIssue(
-                    level="info", category="syntax", message="Test message"
-                )
-                assert issue is not None
-            except ImportError:
-                pytest.skip("ValidationIssue not available")
-
-    def test_validation_issue_creation(self):
-        """Test creating validation issues"""
-        try:
-            # Test issue creation with different severity levels
-            levels = ["info", "warning", "error"]
-
-            for level in levels:
-                try:
-                    issue = ValidationIssue(
-                        level=level,
-                        category="syntax",
-                        message=f"Test {level} issue",
-                        line_number=1,
-                        column_number=1,
-                    )
-                    assert issue is not None
-                    assert issue.level == level
-                except TypeError:
-                    # If constructor signature is different, test basic creation
-                    issue = ValidationIssue(
-                        level=level, category="syntax", message=f"Test {level} message"
-                    )
-                    assert issue is not None
-
-        except ImportError:
-            pytest.skip("ValidationIssue not available")
-
-    def test_validation_issue_formatting(self):
-        """Test validation issue formatting"""
-        try:
-            # Test issue formatting for display
-            try:
-                issue = ValidationIssue(
-                    level="error",
-                    category="syntax",
-                    message="Test formatting issue",
-                    line_number=5,
-                    column_number=10,
-                )
-            except TypeError:
-                issue = ValidationIssue(
-                    level="error", category="syntax", message="Test formatting message"
-                )
-
-            # Test string representation
-            issue_str = str(issue)
-            assert isinstance(issue_str, str)
-            assert len(issue_str) > 0
-
-            # Test if issue contains relevant information
-            if hasattr(issue, "message"):
-                assert issue.message in issue_str
-
-        except ImportError:
-            pytest.skip("ValidationIssue not available")
-
-
-class TestValidationReporter:
-    """Test validation reporter functionality"""
-
-    def test_validation_reporter_initialization(self):
-        """Test ValidationReporter initialization"""
-        try:
-            reporter = ValidationReporter()
-            assert reporter is not None
-        except ImportError:
-            pytest.skip("ValidationReporter not available")
-
-    def test_validation_reporter_issue_collection(self):
-        """Test collecting validation issues"""
-        try:
-            reporter = ValidationReporter()
-
-            # Test issue collection
-            assert reporter is not None
-        except ImportError:
-            pytest.skip("ValidationReporter not available")
-
-    def test_validation_reporter_report_generation(self):
-        """Test generating validation reports"""
-        try:
-            reporter = ValidationReporter()
-
-            # Test report generation
-            assert reporter is not None
-        except ImportError:
-            pytest.skip("ValidationReporter not available")
-
-
-class TestStructureValidator:
-    """Test structure validator functionality"""
-
-    def test_structure_validator_initialization(self):
-        """Test StructureValidator initialization"""
-        try:
-            validator = StructureValidator()
-            assert validator is not None
-        except ImportError:
-            pytest.skip("StructureValidator not available")
-
-    def test_structure_validator_nesting_validation(self):
-        """Test nesting structure validation"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;外側;;; ;;;内側;;; テスト ;;; ;;;")
-            temp_path = f.name
-
-        try:
-            validator = StructureValidator()
-
-            # Test nesting validation
-            assert validator is not None
-            assert Path(temp_path).exists()
-        except ImportError:
-            pytest.skip("StructureValidator not available")
-        finally:
-            os.unlink(temp_path)
-
-    def test_structure_validator_hierarchy_validation(self):
-        """Test hierarchy validation"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write("# 見出し1\n## 見出し2\n### 見出し3")
-            temp_path = f.name
-
-        try:
-            validator = StructureValidator()
-
-            # Test hierarchy validation
-            assert validator is not None
-            assert Path(temp_path).exists()
-        except ImportError:
-            pytest.skip("StructureValidator not available")
-        finally:
-            os.unlink(temp_path)
-
-    def test_structure_validator_balance_validation(self):
-        """Test tag balance validation"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;強調;;; テスト ;;; ;;;強調;;; テスト2 ;;;")
-            temp_path = f.name
-
-        try:
-            validator = StructureValidator()
-
-            # Test balance validation
-            assert validator is not None
-            assert Path(temp_path).exists()
-        except ImportError:
-            pytest.skip("StructureValidator not available")
-        finally:
-            os.unlink(temp_path)
-
-
-class TestSyntaxValidator:
-    """Test syntax validator functionality"""
-
-    def test_syntax_validator_initialization(self):
-        """Test SyntaxValidator initialization"""
-        try:
-            validator = SyntaxValidator()
-            assert validator is not None
-        except ImportError:
-            pytest.skip("SyntaxValidator not available")
-
-    def test_syntax_validator_basic_syntax(self):
-        """Test basic syntax validation"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;強調;;; テスト内容 ;;;")
-            temp_path = f.name
-
-        try:
-            validator = SyntaxValidator()
-
-            # Test basic syntax validation
-            assert validator is not None
-            assert Path(temp_path).exists()
-        except ImportError:
-            pytest.skip("SyntaxValidator not available")
-        finally:
-            os.unlink(temp_path)
-
-    def test_syntax_validator_advanced_syntax(self):
-        """Test advanced syntax validation"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;強調;;; テスト ;;;\n((脚注))\n｜傍注《ぼうちゅう》")
-            temp_path = f.name
-
-        try:
-            validator = SyntaxValidator()
-
-            # Test advanced syntax validation
-            assert validator is not None
-            assert Path(temp_path).exists()
-        except ImportError:
-            pytest.skip("SyntaxValidator not available")
-        finally:
-            os.unlink(temp_path)
-
-    def test_syntax_validator_error_reporting(self):
-        """Test syntax error reporting"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;強調;;; 閉じタグなし")  # Missing closing tag
-            temp_path = f.name
-
-        try:
-            validator = SyntaxValidator()
-
-            # Test error reporting
-            assert validator is not None
-            assert Path(temp_path).exists()
-        except ImportError:
-            pytest.skip("SyntaxValidator not available")
-        finally:
-            os.unlink(temp_path)
-
-
-class TestPerformanceValidator:
-    """Test performance validator functionality"""
-
-    def test_performance_validator_initialization(self):
-        """Test PerformanceValidator initialization"""
-        try:
-            validator = PerformanceValidator()
-            assert validator is not None
-        except ImportError:
-            pytest.skip("PerformanceValidator not available")
-
-    def test_performance_validator_processing_time(self):
-        """Test processing time validation"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write("テスト内容" * 100)  # Create moderately sized content
-            temp_path = f.name
-
-        try:
-            validator = PerformanceValidator()
-
-            # Test processing time validation
-            assert validator is not None
-            assert Path(temp_path).exists()
-        except ImportError:
-            pytest.skip("PerformanceValidator not available")
-        finally:
-            os.unlink(temp_path)
-
-    def test_performance_validator_memory_usage(self):
-        """Test memory usage validation"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write("テスト内容" * 1000)  # Create larger content
-            temp_path = f.name
-
-        try:
-            validator = PerformanceValidator()
-
-            # Test memory usage validation
-            assert validator is not None
-            assert Path(temp_path).exists()
-        except ImportError:
-            pytest.skip("PerformanceValidator not available")
-        finally:
-            os.unlink(temp_path)
-
-    def test_performance_validator_complexity_analysis(self):
-        """Test complexity analysis"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            # Create complex nested structure
-            complex_content = ";;;外側;;; " + ";;;内側;;; テスト ;;; " * 10 + ";;;"
-            f.write(complex_content)
-            temp_path = f.name
-
-        try:
-            validator = PerformanceValidator()
-
-            # Test complexity analysis
-            assert validator is not None
-            assert Path(temp_path).exists()
-        except ImportError:
-            pytest.skip("PerformanceValidator not available")
-        finally:
-            os.unlink(temp_path)
-
-
-class TestValidatorIntegration:
-    """Test validator integration scenarios"""
-
-    def test_validator_integration_end_to_end(self):
-        """Test complete validation workflow"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;強調;;; テスト内容 ;;;\n((脚注内容))\n｜傍注《ぼうちゅう》")
-            temp_path = f.name
-
-        try:
-            # Test complete validation workflow
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
-
-    def test_validator_integration_multiple_files(self):
-        """Test validation with multiple files"""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            files = []
-            for i in range(3):
-                file_path = Path(temp_dir) / f"file_{i}.txt"
-                file_path.write_text(f";;;強調;;; テスト{i} ;;;", encoding="utf-8")
-                files.append(file_path)
-
-            # Test multi-file validation
-            assert all(f.exists() for f in files)
-
-    def test_validator_integration_progressive_validation(self):
-        """Test progressive validation"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;強調;;; テスト内容 ;;;")
-            temp_path = f.name
-
-        try:
-            # Test progressive validation
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
-
-    def test_validator_integration_error_recovery(self):
-        """Test validation error recovery"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;強調;;; エラー1 ;;;\n;;;正常;;; 正常内容 ;;;")
-            temp_path = f.name
-
-        try:
-            # Test error recovery
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
-
-
-class TestValidatorErrorScenarios:
-    """Test validator error scenarios"""
-
-    def test_validator_malformed_syntax(self):
-        """Test handling of malformed syntax"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;強調;;; 閉じタグなし\n;;;; 不正なタグ")
-            temp_path = f.name
-
-        try:
-            # Test malformed syntax handling
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
-
-    def test_validator_encoding_errors(self):
-        """Test handling of encoding errors"""
-        with tempfile.NamedTemporaryFile(mode="wb", delete=False, suffix=".txt") as f:
-            f.write(b"\xff\xfe\x00\x00")  # Invalid UTF-8
-            temp_path = f.name
-
-        try:
-            # Test encoding error handling
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
-
-    def test_validator_circular_references(self):
-        """Test handling of circular references"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;参照;;; {ref1} ;;;\n;;;参照;;; {ref2} ;;;")
-            temp_path = f.name
-
-        try:
-            # Test circular reference handling
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
-
-    def test_validator_memory_overflow(self):
-        """Test handling of memory overflow scenarios"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            # Create very large nested structure
-            large_content = ";;;外側;;; " + ";;;内側;;; テスト ;;; " * 10000 + ";;;"
-            f.write(large_content)
-            temp_path = f.name
-
-        try:
-            # Test memory overflow handling
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
-
-
-class TestValidatorPerformance:
-    """Test validator performance characteristics"""
-
-    def test_validator_performance_large_files(self):
-        """Test validator performance with large files"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            # Create a large file
-            large_content = ";;;強調;;; テスト ;;;\n" * 10000
-            f.write(large_content)
-            temp_path = f.name
-
-        try:
-            # Test large file validation performance
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
-
-    def test_validator_performance_complex_nesting(self):
-        """Test validator performance with complex nesting"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            # Create deeply nested structure
-            nested_content = ";;;外側;;; " * 100 + "テスト" + " ;;;" * 100
-            f.write(nested_content)
-            temp_path = f.name
-
-        try:
-            # Test complex nesting validation performance
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
-
-    def test_validator_performance_concurrent_validation(self):
-        """Test validator performance with concurrent validation"""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            files = []
-            for i in range(10):
-                file_path = Path(temp_dir) / f"concurrent_{i}.txt"
-                file_path.write_text(f";;;強調;;; 並行テスト{i} ;;;", encoding="utf-8")
-                files.append(file_path)
-
-            # Test concurrent validation performance
-            assert all(f.exists() for f in files)
-
-    def test_validator_performance_incremental_validation(self):
-        """Test validator performance with incremental validation"""
-        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
-            f.write(";;;強調;;; 初期内容 ;;;")
-            temp_path = f.name
-
-        try:
-            # Test incremental validation performance
-            assert Path(temp_path).exists()
-
-            # Simulate incremental changes
-            Path(temp_path).write_text(";;;強調;;; 更新内容 ;;;", encoding="utf-8")
-
-            assert Path(temp_path).exists()
-        finally:
-            os.unlink(temp_path)
