+++ tests/test_validators_performance.py
@@ -1,0 +1,291 @@
+"""Validators Performance Tests for Issue 500 Phase 3C
+
+This module tests performance and structure validation functionality.
+"""
+
+import os
+import tempfile
+from pathlib import Path
+from unittest.mock import Mock, patch
+
+import pytest
+
+from kumihan_formatter.core.validators.performance_validator import PerformanceValidator
+from kumihan_formatter.core.validators.structure_validator import StructureValidator
+from tests.test_base import BaseTestCase, ValidatorTestCase, create_test_kumihan_content
+
+
+class TestPerformanceValidator(BaseTestCase):
+    """Test performance validator functionality"""
+
+    def test_performance_validator_initialization(self):
+        """Test PerformanceValidator initialization"""
+        try:
+            validator = PerformanceValidator()
+            assert validator is not None
+
+        except ImportError:
+            pytest.skip("PerformanceValidator not available")
+
+    def test_performance_validator_basic_validation(self):
+        """Test basic performance validation"""
+        try:
+            validator = PerformanceValidator()
+
+            # Test basic validation
+            test_content = ";;;強調;;; テスト内容 ;;;"
+            temp_file = self.create_temp_file(test_content)
+
+            # Test validation
+            if hasattr(validator, "validate"):
+                result = validator.validate(test_content)
+                assert result is not None
+
+            # Test file validation
+            if hasattr(validator, "validate_file"):
+                result = validator.validate_file(temp_file)
+                assert result is not None
+
+        except ImportError:
+            pytest.skip("PerformanceValidator not available")
+
+    def test_performance_validator_large_content(self):
+        """Test performance validation with large content"""
+        try:
+            validator = PerformanceValidator()
+
+            # Test with large content
+            large_content = ";;;強調;;; テスト内容 ;;; " * 1000
+            temp_file = self.create_temp_file(large_content)
+
+            # Test validation performance
+            if hasattr(validator, "validate"):
+                result = validator.validate(large_content)
+                assert result is not None
+
+        except ImportError:
+            pytest.skip("PerformanceValidator not available")
+
+    def test_performance_validator_metrics(self):
+        """Test performance validation metrics"""
+        try:
+            validator = PerformanceValidator()
+
+            # Test metrics collection
+            if hasattr(validator, "get_metrics"):
+                metrics = validator.get_metrics()
+                assert isinstance(metrics, dict)
+
+        except ImportError:
+            pytest.skip("PerformanceValidator not available")
+
+    def test_performance_validator_timing(self):
+        """Test performance timing validation"""
+        try:
+            validator = PerformanceValidator()
+
+            # Test timing metrics
+            test_content = ";;;強調;;; テスト内容 ;;;"
+
+            if hasattr(validator, "validate_with_timing"):
+                result, timing = validator.validate_with_timing(test_content)
+                assert result is not None
+                assert timing >= 0
+
+        except ImportError:
+            pytest.skip("PerformanceValidator not available")
+
+    def test_performance_validator_memory_usage(self):
+        """Test performance memory usage validation"""
+        try:
+            validator = PerformanceValidator()
+
+            # Test memory usage
+            if hasattr(validator, "get_memory_usage"):
+                memory_usage = validator.get_memory_usage()
+                assert memory_usage >= 0
+
+        except ImportError:
+            pytest.skip("PerformanceValidator not available")
+
+    def test_performance_validator_benchmarks(self):
+        """Test performance benchmarking"""
+        try:
+            validator = PerformanceValidator()
+
+            # Test benchmarking
+            if hasattr(validator, "run_benchmark"):
+                benchmark_result = validator.run_benchmark()
+                assert benchmark_result is not None
+
+        except ImportError:
+            pytest.skip("PerformanceValidator not available")
+
+
+class TestStructureValidator(BaseTestCase):
+    """Test structure validator functionality"""
+
+    def test_structure_validator_initialization(self):
+        """Test StructureValidator initialization"""
+        try:
+            validator = StructureValidator()
+            assert validator is not None
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_basic_validation(self):
+        """Test basic structure validation"""
+        try:
+            validator = StructureValidator()
+
+            # Test basic validation
+            test_content = ";;;強調;;; テスト内容 ;;;"
+            temp_file = self.create_temp_file(test_content)
+
+            # Test validation
+            if hasattr(validator, "validate"):
+                result = validator.validate(test_content)
+                assert result is not None
+
+            # Test file validation
+            if hasattr(validator, "validate_file"):
+                result = validator.validate_file(temp_file)
+                assert result is not None
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_nested_structures(self):
+        """Test structure validation with nested structures"""
+        try:
+            validator = StructureValidator()
+
+            # Test nested structures
+            nested_content = ";;;強調;;; ;;;傍注;;; ネスト ;;; ;;;"
+            temp_file = self.create_temp_file(nested_content)
+
+            # Test validation
+            if hasattr(validator, "validate"):
+                result = validator.validate(nested_content)
+                assert result is not None
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_error_detection(self):
+        """Test structure error detection"""
+        try:
+            validator = StructureValidator()
+
+            # Test error detection
+            malformed_content = ";;;強調;;; 未完成構造"
+            temp_file = self.create_temp_file(malformed_content)
+
+            # Test validation
+            if hasattr(validator, "validate"):
+                result = validator.validate(malformed_content)
+                assert result is not None
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_balanced_tags(self):
+        """Test balanced tag validation"""
+        try:
+            validator = StructureValidator()
+
+            # Test balanced tags
+            balanced_content = ";;;強調;;; テスト ;;;"
+            unbalanced_content = ";;;強調;;; テスト"
+
+            if hasattr(validator, "validate_balanced_tags"):
+                balanced_result = validator.validate_balanced_tags(balanced_content)
+                unbalanced_result = validator.validate_balanced_tags(unbalanced_content)
+
+                assert balanced_result != unbalanced_result
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_hierarchy(self):
+        """Test structure hierarchy validation"""
+        try:
+            validator = StructureValidator()
+
+            # Test hierarchy validation
+            if hasattr(validator, "validate_hierarchy"):
+                hierarchy_result = validator.validate_hierarchy(";;;強調;;; テスト ;;;")
+                assert hierarchy_result is not None
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_depth_analysis(self):
+        """Test structure depth analysis"""
+        try:
+            validator = StructureValidator()
+
+            # Test depth analysis
+            deep_content = ";;;強調;;; ;;;傍注;;; ;;;内側;;; テスト ;;; ;;; ;;;"
+
+            if hasattr(validator, "analyze_depth"):
+                depth_result = validator.analyze_depth(deep_content)
+                assert depth_result is not None
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_complexity_metrics(self):
+        """Test structure complexity metrics"""
+        try:
+            validator = StructureValidator()
+
+            # Test complexity metrics
+            if hasattr(validator, "get_complexity_metrics"):
+                metrics = validator.get_complexity_metrics()
+                assert isinstance(metrics, dict)
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_pattern_analysis(self):
+        """Test structure pattern analysis"""
+        try:
+            validator = StructureValidator()
+
+            # Test pattern analysis
+            if hasattr(validator, "analyze_patterns"):
+                patterns = validator.analyze_patterns(";;;強調;;; テスト ;;;")
+                assert patterns is not None
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_consistency_check(self):
+        """Test structure consistency checking"""
+        try:
+            validator = StructureValidator()
+
+            # Test consistency checking
+            if hasattr(validator, "check_consistency"):
+                consistency_result = validator.check_consistency(
+                    ";;;強調;;; テスト ;;;"
+                )
+                assert consistency_result is not None
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
+
+    def test_structure_validator_repair_suggestions(self):
+        """Test structure repair suggestions"""
+        try:
+            validator = StructureValidator()
+
+            # Test repair suggestions
+            if hasattr(validator, "suggest_repairs"):
+                suggestions = validator.suggest_repairs(";;;強調;;; 不完全")
+                assert suggestions is not None
+
+        except ImportError:
+            pytest.skip("StructureValidator not available")
